Redesigned testing structure
