+++ tests/altr_files/3multi_var.altr
