+++ tests/altr_files/4multi_letter_var.altr
