+++ tests/altr_files/integration.altr
@@ -1,0 +1,7 @@
+a = 1
+b = false
+Name = "Alter"
+
+say "Hello!"
+say a
+say a,b,Name
